+++ Perjalanan.java
@@ -1,10 +1,10 @@
 public class Perjalanan{
-    private long id;
+    private int id;
     private double jarakKm;
     private String tanggal;
     private Kendaraan kendaraan;
 
-    public Perjalanan(long id, double jarakKm, String tanggal, Kendaraan kendaraan) {
+    public Perjalanan(int id, double jarakKm, String tanggal, Kendaraan kendaraan) {
         this.id = id;
         this.jarakKm = jarakKm;
         this.tanggal = tanggal;
